Hoist shared docker test fakes into sandbox conftest
- Move _FakeBundleStore, _RecordingDocker and _flag_values into
  apps/worker/tests/sandbox/conftest.py (superset test_docker.py versions)
- Import them from both test_docker.py and test_vector_credential_forwarding.py
  instead of re-defining, mirroring the existing .conftest import in
  test_substrate.py
- Replace the vector lane's _forwarded_names helper with
  [v for v in _flag_values(argv, "-e") if "=" not in v], keeping the
  -e NAME vs -e KEY=VALUE distinction as the only new logic
- Skip the optional parents[4] repo-root centralization (new abstraction,
  not reuse)

Ref #604
