+++ apps/worker/tests/sandbox/test_docker.py
@@ -20,28 +20,7 @@ from agentos_worker.sandbox.docker import (
     RunnerHardening,
 )
 
-
-class _FakeBundleStore:
-    def __init__(self, data: bytes = b"") -> None:
-        self._data = data
-        self.requested: list[str] = []
-
-    def get(self, key: str) -> bytes:
-        self.requested.append(key)
-        return self._data
-
-
-class _RecordingDocker(DockerSandboxClient):
-    """Captures every docker argv and returns canned stdout per subcommand."""
-
-    def __init__(self, **kwargs: object) -> None:
-        super().__init__(**kwargs)  # type: ignore[arg-type]
-        self.calls: list[list[str]] = []
-        self.outputs: dict[str, str] = {}
-
-    def _docker(self, args: list[str], *, check: bool = True) -> str:
-        self.calls.append(args)
-        return self.outputs.get(args[0], "")
+from .conftest import _FakeBundleStore, _flag_values, _RecordingDocker
 
 
 def _plugin_tar_gz(wrapper: str | None) -> bytes:
@@ -55,10 +34,6 @@ def _plugin_tar_gz(wrapper: str | None) -> bytes:
         info.size = len(manifest)
         tf.addfile(info, io.BytesIO(manifest))
     return buf.getvalue()
-
-
-def _flag_values(argv: list[str], flag: str) -> list[str]:
-    return [argv[i + 1] for i, a in enumerate(argv) if a == flag and i + 1 < len(argv)]
 
 
 def test_create_claim_argv_carries_boot_env() -> None:
